refactor(web): split api/client.ts into http plumbing and endpoint modules
client.ts was one 1759-line file: the fetch helpers plus a single `api`
object holding every endpoint of the dashboard. It now keeps only the
public surface (the helper re-exports and the assembled `api` object,
50 lines) and delegates to:

- api/http.ts: ApiError, fetchOK/fetchJSON, CSRF and auth-header helpers
- api/endpoints/{auth,providers,models,logs,settings,users,chat}.ts: the
  endpoint groups, one module per dashboard area, each exporting its
  groups by their original key names

Pure moves; `api.<group>.<call>` keeps the same names and types, so the
59 importers and the 17 client.*.test.ts suites are untouched. The size
allowlist drops the client.ts entry.
